+++ NEWS.md
@@ -1,3 +1,8 @@
+# psychTestR 2.13.3
+
+- psychTestR now throws an error on nested calls to `new_timeline()`,
+which sometimes worked as intended, sometimes not...
+
 # psychTestR 2.13.2
 
 - Fixed issue where `c()` didn't work well for all combinations